Upgraded the whole project to using Gulp 4
I upgraded by replacing all the gulp files with the gulp files from the latest version of Yeogurt then tweacked them to  fit my needs.
