+++ docs/index.html
@@ -27,7 +27,7 @@
   <meta name="twitter:description" content="An accessible polyfill for `&lt;input type='time'/&gt;` elements modeled after the Chrome &amp; Firefox desktop implementations.">
   <meta name="twitter:image" content="https://dan503.github.io/time-input-polyfill/images/og-image.jpg">
   <meta name="twitter:image:alt" content="time-input-polyfill logo against a swirling background">
-  <link rel="stylesheet" href="./styles/main.css">
+  <link rel="stylesheet" href="./styles/main-68b1d7d742.css">
   <!-- Google Analytics-->
   <script async src="https://www.googletagmanager.com/gtag/js?id=UA-90397907-4"></script>
   <script>
@@ -44,7 +44,7 @@
   </div>
   <div class="container" style="text-align: center">
     <h1>Time Input Polyfill</h1>
-    <p class="version">v1.0.7</p>
+    <p class="version">v1.0.8</p>
         <p class="forkMe"><a class="forkMe__link" href="https://github.com/Dan503/time-input-polyfill">Fork me on GitHub</a></p>
         <div class="textBlock">
           <p>The demos on this page have been applied manually in order to demonstrate the difference between a real time input and a polyfill time input in modern browsers.</p>
@@ -83,8 +83,8 @@
         </div>
         <div class="downloadMethod">
           <h2>Download via CDN:</h2>
-          <pre><code>&lt;script src="https://cdn.jsdelivr.net/npm/time-input-polyfill@1.0.7"&gt;&lt;/script&gt;</code></pre>
-          <p><a href="https://cdn.jsdelivr.net/npm/time-input-polyfill@1.0.7">https://cdn.jsdelivr.net/npm/time-input-polyfill@1.0.7</a></p>
+          <pre><code>&lt;script src="https://cdn.jsdelivr.net/npm/time-input-polyfill@1.0.8"&gt;&lt;/script&gt;</code></pre>
+          <p><a href="https://cdn.jsdelivr.net/npm/time-input-polyfill@1.0.8">https://cdn.jsdelivr.net/npm/time-input-polyfill@1.0.8</a></p>
         </div>
         <div class="downloadMethod">
           <h2>View documentation:</h2>
