+++ docs/rev-manifest.json
@@ -1,0 +1,5 @@
+{
+  "scripts/time-input-polyfill.auto.min.js": "scripts/time-input-polyfill-a974aa3df5.auto.min.js",
+  "scripts/time-input-polyfill.min.js": "scripts/time-input-polyfill-ef316eb7f0.min.js",
+  "styles/main.css": "styles/main-68b1d7d742.css"
+}
