+++ docs/styles/main-68b1d7d742.css
@@ -1,0 +1,1 @@
+/*! normalize.css v8.0.1 | MIT License | github.com/necolas/normalize.css */@import url("https://fonts.googleapis.com/css?family=Teko");html{line-height:1.15;-webkit-text-size-adjust:100%}body{margin:0}main{display:block}h1{font-size:2em;margin:.67em 0}hr{-webkit-box-sizing:content-box;box-sizing:content-box;height:0;overflow:visible}pre{font-family:monospace,monospace;font-size:1em}a{background-color:transparent}abbr[title]{border-bottom:none;text-decoration:underline;-webkit-text-decoration:underline dotted;text-decoration:underline dotted}b,strong{font-weight:bolder}code,kbd,samp{font-family:monospace,monospace;font-size:1em}small{font-size:80%}sub,sup{font-size:75%;line-height:0;position:relative;vertical-align:baseline}sub{bottom:-.25em}sup{top:-.5em}img{border-style:none}button,input,optgroup,select,textarea{font-family:inherit;font-size:100%;line-height:1.15;margin:0}button,input{overflow:visible}button,select{text-transform:none}[type=button],[type=reset],[type=submit],button{-webkit-appearance:button}[type=button]::-moz-focus-inner,[type=reset]::-moz-focus-inner,[type=submit]::-moz-focus-inner,button::-moz-focus-inner{border-style:none;padding:0}[type=button]:-moz-focusring,[type=reset]:-moz-focusring,[type=submit]:-moz-focusring,button:-moz-focusring{outline:1px dotted ButtonText}fieldset{padding:.35em .75em .625em}legend{-webkit-box-sizing:border-box;box-sizing:border-box;color:inherit;display:table;max-width:100%;padding:0;white-space:normal}progress{vertical-align:baseline}textarea{overflow:auto}[type=checkbox],[type=radio]{-webkit-box-sizing:border-box;box-sizing:border-box;padding:0}[type=number]::-webkit-inner-spin-button,[type=number]::-webkit-outer-spin-button{height:auto}[type=search]{-webkit-appearance:textfield;outline-offset:-2px}[type=search]::-webkit-search-decoration{-webkit-appearance:none}::-webkit-file-upload-button{-webkit-appearance:button;font:inherit}details{display:block}summary{display:list-item}[hidden],template{display:none}*,:after,:before{-webkit-box-sizing:border-box;box-sizing:border-box}a{color:#fff;word-wrap:break-word}a:focus,a:hover{text-decoration:none}html{background:#292929;color:#c3c3c3;font-size:.625em;font-family:sans-serif}body{font-size:3rem;text-align:center}@media (max-width:600px){body{font-size:2rem}}.container{padding:50px 30px;position:relative;z-index:1}@media (max-width:600px){.container{padding-top:10rem}}h1{margin-bottom:.1em}input{color:#292929;background-color:#c3c3c3;border:1px solid #000;padding:10px;max-width:100%;text-align:left;text-align:start}button{color:#fff;background:#000;padding:10px 20px;border:2px solid;border-radius:10px;-webkit-transition:.3s;transition:.3s}button:focus,button:hover{background:#fff;color:#000}.block{display:-webkit-box;display:-ms-flexbox;display:flex;-webkit-box-orient:vertical;-webkit-box-direction:normal;-ms-flex-direction:column;flex-direction:column;-webkit-box-align:center;-ms-flex-align:center;align-items:center;margin-top:1.5em}.block label{display:block;margin-bottom:.5em}pre{margin:.5em 0;overflow:visible;white-space:pre-wrap}pre code{font-size:.6em;background:#000;padding:.8em 20px;border-radius:10px;-webkit-box-decoration-break:clone;box-decoration-break:clone;line-height:2;word-wrap:break-word}.credit{font-size:1.2rem;padding:0 20px}.version{font-size:2rem;margin-top:0;font-style:italic}.fork-message{margin:0;position:absolute;top:0;right:0}.fork-message a,.fork-message img{display:block}.clock{position:fixed;top:0;left:0;right:0;bottom:0;margin:auto;height:90vmin;width:90vmin;opacity:.3}.visually-hidden{position:absolute;height:0;width:0;overflow:hidden;opacity:0;pointer-events:none}.downloadMethod{text-align:center;margin:2em}.downloadMethod h2{font-size:1em;margin-bottom:0}.downloadMethod p{margin-top:0}.downloadMethod a{font-size:.7em}.result{display:none;background:rgba(0,0,0,.5);border-radius:50px;max-width:47rem;padding:50px;margin:auto;margin-top:2em;outline:none}.result.-visible{display:block}.result__heading{margin-top:0;padding-bottom:1rem;border-bottom:1px solid #fff}.result__list{margin-bottom:0}.result__item{display:-webkit-box;display:-ms-flexbox;display:flex;-webkit-box-pack:end;-ms-flex-pack:end;justify-content:flex-end;-webkit-box-align:center;-ms-flex-align:center;align-items:center}.result__item+.result__item{margin-top:2rem}.result__label{color:#fff;text-align:right}.result__value{margin-left:1rem;min-width:7ch;-ms-flex-negative:0;flex-shrink:0;line-height:1}.result__close{margin-top:1.5em}.textBlock{background:rgba(0,0,0,.5);border-radius:50px;max-width:60rem;padding:40px;margin:auto;margin-top:2em;font-size:.9em;line-height:1.5}.textBlock p{margin:0}.textBlock p+p{margin-top:1.5em}.forkMe{margin:0;position:absolute;top:0;right:0;padding-top:11.5rem;overflow:hidden}@media (min-width:600px){.forkMe{position:fixed}}@media (min-width:1200px){.forkMe{padding-top:16.5rem}}.forkMe__link{display:block;background:#fff;color:#292929;font-size:2.2rem;line-height:1;font-family:Teko,sans-serif;position:relative;padding:10px 4rem;margin-right:-1rem;width:23rem;-webkit-transform-origin:bottom right;transform-origin:bottom right;-webkit-transform:rotate(45deg);transform:rotate(45deg);text-decoration:none;text-align:center;-webkit-box-shadow:0 0 20px #000;box-shadow:0 0 20px #000;-webkit-transition:.2s;transition:.2s}@media (min-width:1200px){.forkMe__link{font-size:3rem;width:31rem}}.forkMe__link:after,.forkMe__link:before{content:"";border-bottom:1px dotted #979797;height:0;width:100%;position:absolute;left:0}.forkMe__link:before{top:2px}.forkMe__link:after{bottom:2px}.forkMe__link:focus,.forkMe__link:hover{background:#f7ff85;color:#292929}
