--- docs/styles/main.css
@@ -1,2 +1,0 @@
-/*! normalize.css v3.0.3 | MIT License | github.com/necolas/normalize.css */@import url("https://fonts.googleapis.com/css?family=Teko");html{font-family:sans-serif;-ms-text-size-adjust:100%;-webkit-text-size-adjust:100%}body{margin:0}article,aside,details,figcaption,figure,footer,header,hgroup,main,menu,nav,section,summary{display:block}audio,canvas,progress,video{display:inline-block;vertical-align:baseline}audio:not([controls]){display:none;height:0}[hidden],template{display:none}a{background-color:transparent}a:active,a:hover{outline:0}abbr[title]{border-bottom:1px dotted}b,strong{font-weight:700}dfn{font-style:italic}h1{font-size:2em;margin:.67em 0}mark{background:#ff0;color:#000}small{font-size:80%}sub,sup{font-size:75%;line-height:0;position:relative;vertical-align:baseline}sup{top:-.5em}sub{bottom:-.25em}img{border:0}svg:not(:root){overflow:hidden}figure{margin:1em 40px}hr{box-sizing:content-box;height:0}pre{overflow:auto}code,kbd,pre,samp{font-family:monospace,monospace;font-size:1em}button,input,optgroup,select,textarea{color:inherit;font:inherit;margin:0}button{overflow:visible}button,select{text-transform:none}button,html input[type=button],input[type=reset],input[type=submit]{-webkit-appearance:button;cursor:pointer}button[disabled],html input[disabled]{cursor:default}button::-moz-focus-inner,input::-moz-focus-inner{border:0;padding:0}input{line-height:normal}input[type=checkbox],input[type=radio]{box-sizing:border-box;padding:0}input[type=number]::-webkit-inner-spin-button,input[type=number]::-webkit-outer-spin-button{height:auto}input[type=search]{-webkit-appearance:textfield;box-sizing:content-box}input[type=search]::-webkit-search-cancel-button,input[type=search]::-webkit-search-decoration{-webkit-appearance:none}fieldset{border:1px solid silver;margin:0 2px;padding:.35em .625em .75em}legend{border:0;padding:0}textarea{overflow:auto}optgroup{font-weight:700}table{border-collapse:collapse;border-spacing:0}td,th{padding:0}*,:after,:before{box-sizing:border-box}a{color:#fff;word-wrap:break-word}a:focus,a:hover{text-decoration:none}html{background:#292929;color:#c3c3c3;font-size:.625em}body{font-size:3rem;text-align:center}@media (max-width:600px){body{font-size:2rem}}.container{padding:50px 30px;position:relative;z-index:1}@media (max-width:600px){.container{padding-top:10rem}}h1{margin-bottom:.1em}input{color:#292929;background-color:#c3c3c3;border:1px solid #000;padding:10px;max-width:100%;text-align:left;text-align:start}button{color:#fff;background:#000;padding:10px 20px;border:2px solid;border-radius:10px;transition:.3s}button:focus,button:hover{background:#fff;color:#000}.block{display:-ms-flexbox;display:flex;-ms-flex-direction:column;flex-direction:column;-ms-flex-align:center;align-items:center;margin-top:1.5em}.block label{display:block;margin-bottom:.5em}pre{margin:.5em 0;overflow:visible;white-space:pre-wrap}pre code{font-size:.6em;background:#000;padding:.8em 20px;border-radius:10px;-webkit-box-decoration-break:clone;box-decoration-break:clone;line-height:2;word-wrap:break-word}.credit{font-size:1.2rem;padding:0 20px}.version{font-size:2rem;margin-top:0;font-style:italic}.fork-message{margin:0;position:absolute;top:0;right:0}.fork-message a,.fork-message img{display:block}.clock{position:fixed;top:0;left:0;right:0;bottom:0;margin:auto;height:90vmin;width:90vmin;opacity:.3}.visually-hidden{position:absolute;height:0;width:0;overflow:hidden;opacity:0;pointer-events:none}.downloadMethod{text-align:center;margin:2em}.downloadMethod h2{font-size:1em;margin-bottom:0}.downloadMethod p{margin-top:0}.downloadMethod a{font-size:.7em}.result{display:none;background:rgba(0,0,0,.5);border-radius:50px;max-width:47rem;padding:50px;margin:auto;margin-top:2em;outline:none}.result.-visible{display:block}.result__heading{margin-top:0;padding-bottom:1rem;border-bottom:1px solid #fff}.result__list{margin-bottom:0}.result__item{display:-ms-flexbox;display:flex;-ms-flex-pack:end;justify-content:flex-end;-ms-flex-align:center;align-items:center}.result__item+.result__item{margin-top:2rem}.result__label{color:#fff;text-align:right}.result__value{margin-left:1rem;min-width:7ch;-ms-flex-negative:0;flex-shrink:0;line-height:1}.result__close{margin-top:1.5em}.textBlock{background:rgba(0,0,0,.5);border-radius:50px;max-width:60rem;padding:40px;margin:auto;margin-top:2em;font-size:.9em;line-height:1.5}.textBlock p{margin:0}.textBlock p+p{margin-top:1.5em}.forkMe{margin:0;position:absolute;top:0;right:0;padding-top:11.5rem;overflow:hidden}@media (min-width:600px){.forkMe{position:fixed}}@media (min-width:1200px){.forkMe{padding-top:16.5rem}}.forkMe__link{display:block;background:#fff;color:#292929;font-size:2.2rem;line-height:1;font-family:Teko,sans-serif;position:relative;padding:10px 4rem;margin-right:-1rem;width:23rem;transform-origin:bottom right;transform:rotate(45deg);text-decoration:none;text-align:center;box-shadow:0 0 20px #000;transition:.2s}@media (min-width:1200px){.forkMe__link{font-size:3rem;width:31rem}}.forkMe__link:after,.forkMe__link:before{content:"";border-bottom:1px dotted #979797;height:0;width:100%;position:absolute;left:0}.forkMe__link:before{top:2px}.forkMe__link:after{bottom:2px}.forkMe__link:focus,.forkMe__link:hover{background:#f7ff85;color:#292929}
-/*# sourceMappingURL=main.css.map */
