--- gulp/clean.js
@@ -1,15 +1,0 @@
-'use strict'
-
-var del = require('del')
-
-module.exports = function(
-	gulp,
-	plugins,
-	args,
-	config,
-	taskTarget,
-	browserSync,
-) {
-	// Clean
-	gulp.task('clean', del.bind(null, [taskTarget]))
-}
