+++ package.json
@@ -1,96 +1,120 @@
 {
-  "name": "time-input-polyfill",
-  "description": "An accessible polyfill for `<input type='time'/>` elements modeled after the Chrome & Firefox desktop implementations.",
-  "version": "1.0.8",
-  "main": "index.js",
-  "jsdelivr": "dist/time-input-polyfill.auto.min.js",
-  "dependencies": {},
-  "devDependencies": {
-    "autoprefixer": "^9.1.5",
-    "browser-sync": "~2.11.1",
-    "browserify": "~13.0.0",
-    "del": "~2.2.0",
-    "envify": "~3.4.0",
-    "foldero": "~0.1.1",
-    "glob": "~6.0.4",
-    "gulp": "~3.9.0",
-    "gulp-changed": "~1.3.0",
-    "gulp-cssnano": "~2.1.0",
-    "gulp-data": "~1.2.1",
-    "gulp-eslint": "~1.1.1",
-    "gulp-file-loader": "^2.0.1",
-    "gulp-header": "^2.0.5",
-    "gulp-htmlmin": "~1.3.0",
-    "gulp-if": "~2.0.0",
-    "gulp-imagemin": "~2.4.0",
-    "gulp-load-plugins": "~1.2.0",
-    "gulp-newer": "~1.1.0",
-    "gulp-plumber": "~1.0.1",
-    "gulp-postcss": "^8.0.0",
-    "gulp-pug": "^4.0.1",
-    "gulp-rename": "~1.2.2",
-    "gulp-sass": "^4.0.1",
-    "gulp-sourcemaps": "~1.6.0",
-    "gulp-uglify": "~1.5.1",
-    "gulp-util": "~3.0.7",
-    "imagemin-pngquant": "~4.2.0",
-    "imagemin-svgo": "~4.2.0",
-    "js-yaml": "~3.5.2",
-    "lodash": "~4.0.0",
-    "minimist": "~1.2.0",
-    "normalize.css": "~3.0.3",
-    "rimraf": "~2.5.0",
-    "vinyl-buffer": "~1.0.0",
-    "vinyl-source-stream": "~1.1.0",
-    "watchify": "~3.7.0 ",
-    "wrench": "~1.5.8"
-  },
-  "keywords": [
-    "time",
-    "input",
-    "polyfill",
-    "IE",
-    "accessible",
-    "accessibility",
-    "a11y"
-  ],
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/Dan503/time-input-polyfill.git"
-  },
-  "homepage": "https://dan503.github.io/time-input-polyfill/",
-  "author": "Daniel Tonon",
-  "license": "MIT",
-  "bugs": {
-    "url": "https://github.com/Dan503/time-input-polyfill/issues"
-  },
-  "scripts": {
-    "clean-deps": "rimraf node_modules"
-  },
-  "engines": {
-    "node": ">=4.0.0"
-  },
-  "//": "CUSTOM CONFIGURATION",
-  "config": {
-    "//": "Entry files",
-    "host": "127.0.0.1",
-    "port": "3000",
-    "baseUrl": "./",
-    "directories": {
-      "source": "src",
-      "destination": "docs",
-      "temporary": "tmp",
-      "//": "Directories relative to `source` directory",
-      "modules": "_modules",
-      "layouts": "_layouts",
-      "images": "_images",
-      "styles": "_styles",
-      "scripts": "_scripts",
-      "data": "_data"
-    },
-    "entries": {
-      "js": "main**.js",
-      "css": "main**.{sass,scss}"
-    }
-  }
+	"name": "time-input-polyfill",
+	"description": "An accessible polyfill for `<input type='time'/>` elements modeled after the Chrome & Firefox desktop implementations.",
+	"version": "1.0.8",
+	"main": "index.js",
+	"jsdelivr": "dist/time-input-polyfill.auto.min.js",
+	"dependencies": {},
+	"devDependencies": {
+		"@babel/core": "~7.0.0",
+		"@babel/preset-env": "~7.0.0",
+		"@babel/register": "~7.4.4",
+		"autoprefixer": "~9.6.1",
+		"babel-eslint": "~9.0.0",
+		"babelify": "~10.0.0",
+		"browser-sync": "~2.26.7",
+		"browserify": "~16.3.0",
+		"del": "~2.2.2",
+		"envify": "~4.1.0",
+		"eslint": "6.1.0",
+		"eslint-config-prettier": "~6.0.0",
+		"foldero": "~0.1.1",
+		"glob": "~7.1.4",
+		"gulp": "~4.0.2",
+		"gulp-changed": "~4.0.0",
+		"gulp-cssnano": "~2.1.3",
+		"gulp-data": "~1.3.1",
+		"gulp-eslint": "~6.0.0",
+		"gulp-filter": "6.0.0",
+		"gulp-header": "^2.0.9",
+		"gulp-htmlmin": "~5.0.1",
+		"gulp-if": "~2.0.2",
+		"gulp-imagemin": "~6.0.0",
+		"gulp-load-plugins": "~2.0.0",
+		"gulp-plumber": "~1.2.1",
+		"gulp-postcss": "~8.0.0",
+		"gulp-pug": "~4.0.1",
+		"gulp-rename": "~1.4.0",
+		"gulp-rev": "9.0.0",
+		"gulp-rev-delete-original": "0.2.3",
+		"gulp-rev-rewrite": "2.0.0",
+		"gulp-sass": "~4.0.2",
+		"gulp-sourcemaps": "~2.6.5",
+		"gulp-terser": "^1.2.0",
+		"gulp-uglify": "~3.0.2",
+		"gulp-util": "~3.0.8",
+		"imagemin-pngquant": "~7.0.0",
+		"imagemin-svgo": "~7.0.0",
+		"js-yaml": "~3.13.0",
+		"lodash": "~4.17.4",
+		"minimist": "~1.2.0",
+		"normalize.css": "~8.0.1",
+		"prettier": "~1.16.4",
+		"pug": "~2.0.4",
+		"rimraf": "2.6.3",
+		"vinyl-buffer": "~1.0.1",
+		"vinyl-source-stream": "~2.0.0",
+		"watchify": "~3.11.0 "
+	},
+	"keywords": [
+		"time",
+		"input",
+		"polyfill",
+		"IE",
+		"accessible",
+		"accessibility",
+		"a11y"
+	],
+	"repository": {
+		"type": "git",
+		"url": "git+https://github.com/Dan503/time-input-polyfill.git"
+	},
+	"homepage": "https://dan503.github.io/time-input-polyfill/",
+	"author": "Daniel Tonon",
+	"license": "MIT",
+	"bugs": {
+		"url": "https://github.com/Dan503/time-input-polyfill/issues"
+	},
+	"scripts": {
+		"clean-deps": "npx rimraf node_modules",
+		"clean": "npx gulp clean",
+		"lint": "npx gulp eslint",
+		"test": "npx gulp test",
+		"format": "npx prettier --single-quote --write src/**/*.js",
+		"serve": "NODE_ENV=development npx gulp serve",
+		"serve:prod": "NODE_ENV=production npx gulp serve --production",
+		"build": "NODE_ENV=production npx gulp --production"
+	},
+	"engines": {
+		"node": ">=8.16.0"
+	},
+	"browserslist": [
+		"last 2 version",
+		"> 5%",
+		"safari 5",
+		"ios 6",
+		"android 4"
+	],
+	"config": {
+		"//": "Entry files",
+		"host": "127.0.0.1",
+		"port": "3000",
+		"baseUrl": "./",
+		"directories": {
+			"source": "src",
+			"destination": "docs",
+			"temporary": "tmp",
+			"//": "Directories relative to `source` directory",
+			"modules": "_modules",
+			"layouts": "_layouts",
+			"images": "_images",
+			"styles": "_styles",
+			"scripts": "_scripts",
+			"data": "_data"
+		},
+		"entries": {
+			"js": "main**.js",
+			"css": "main**.{sass,scss}"
+		}
+	}
 }
