+++ src/_styles/main.scss
@@ -6,140 +6,144 @@
 *,
 *:before,
 *:after {
-  box-sizing: border-box;
+	box-sizing: border-box;
 }
 
 a {
-  color: #fff;
-  word-wrap: break-word;
+	color: #fff;
+	word-wrap: break-word;
 
-  &:hover, &:focus {
-    text-decoration: none;
-  }
+	&:hover,
+	&:focus {
+		text-decoration: none;
+	}
 }
 
 html {
-  background: #292929;
-  color: #c3c3c3;
-  font-size: 0.625em;
+	background: #292929;
+	color: #c3c3c3;
+	font-size: 0.625em;
+	font-family: sans-serif;
 }
 
 body {
-  font-size: 3rem;
-  text-align: center;
+	font-size: 3rem;
+	text-align: center;
 
-  @media (max-width: 600px) {
-    font-size: 2rem;
-  }
+	@media (max-width: 600px) {
+		font-size: 2rem;
+	}
 }
 
 .container {
-  padding: 50px 30px;
-  position: relative;
-  z-index: 1;
+	padding: 50px 30px;
+	position: relative;
+	z-index: 1;
 
-  @media (max-width: 600px) {
-    padding-top: 10rem;
-  }
+	@media (max-width: 600px) {
+		padding-top: 10rem;
+	}
 }
 
 h1 {
-  margin-bottom: 0.1em;
+	margin-bottom: 0.1em;
 }
 
 input {
-  color: #292929;
-  background-color: #c3c3c3;
-  border: 1px solid #000;
-  padding: 10px;
-  max-width: 100%;
-  text-align: left;
-  text-align: start;
+	color: #292929;
+	background-color: #c3c3c3;
+	border: 1px solid #000;
+	padding: 10px;
+	max-width: 100%;
+	text-align: left;
+	text-align: start;
 }
 
 button {
-  color: #fff;
-  background: #000;
-  padding: 10px 20px;
-  border: 2px solid;
-  border-radius: 10px;
-  transition: 0.3s;
+	color: #fff;
+	background: #000;
+	padding: 10px 20px;
+	border: 2px solid;
+	border-radius: 10px;
+	transition: 0.3s;
 
-  &:hover, &:focus {
-    background: #fff;
-    color: #000;
-  }
+	&:hover,
+	&:focus {
+		background: #fff;
+		color: #000;
+	}
 }
 
 .block {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  margin-top: 1.5em;
+	display: flex;
+	flex-direction: column;
+	align-items: center;
+	margin-top: 1.5em;
 
-  label {
-    display: block;
-    margin-bottom: 0.5em;
-  }
+	label {
+		display: block;
+		margin-bottom: 0.5em;
+	}
 }
 
 pre {
-  margin: 0.5em 0;
-  overflow: visible;
-  white-space: pre-wrap;
+	margin: 0.5em 0;
+	overflow: visible;
+	white-space: pre-wrap;
 
-  code {
-    font-size: 0.6em;
-    background: #000;
-    padding: 0.8em 20px;
-    border-radius: 10px;
-    box-decoration-break: clone;
-    line-height: 2;
-    word-wrap: break-word;
-  }
+	code {
+		font-size: 0.6em;
+		background: #000;
+		padding: 0.8em 20px;
+		border-radius: 10px;
+		box-decoration-break: clone;
+		line-height: 2;
+		word-wrap: break-word;
+	}
 }
 
 .credit {
-  font-size: 1.2rem;
-  padding: 0 20px;
+	font-size: 1.2rem;
+	padding: 0 20px;
 }
 
 .version {
-  font-size: 2rem;
-  margin-top: 0;
-  font-style: italic;
+	font-size: 2rem;
+	margin-top: 0;
+	font-style: italic;
 }
 
 .fork-message {
-  margin: 0;
-  position: absolute;
-  top: 0;
-  right: 0;
+	margin: 0;
+	position: absolute;
+	top: 0;
+	right: 0;
 
-  a, img {
-    display: block;
-  }
+	a,
+	img {
+		display: block;
+	}
 }
 
 .clock {
-  position: fixed;
-  top: 0;
-  left: 0;
-  right: 0;
-  bottom: 0;
-  margin: auto;
-  height: 90vmin;
-  width: 90vmin;
-  opacity: 0.3;
+	position: fixed;
+	top: 0;
+	left: 0;
+	right: 0;
+	bottom: 0;
+	margin: auto;
+	height: 90vmin;
+	width: 90vmin;
+	opacity: 0.3;
 }
 
 .visually-hidden {
-  position: absolute;
-  height: 0;
-  width: 0;
-  overflow: hidden;
-  opacity: 0;
-  pointer-events: none;
+	position: absolute;
+	height: 0;
+	width: 0;
+	overflow: hidden;
+	opacity: 0;
+	pointer-events: none;
 }
 
 // Import Modules
